Move service creation to a separate file
